backend/requirements.txt
streamlit
flask
flask-cors
langchain
langchain-community
langchain-core
langchain-huggingface
langchain_mistralai
langchain-google-genai
faiss-cpu
sentence-transformers
unstructured
pypdf
python-docx
tqdm
pdfminer-six
Pillow
python-dotenv
paddlepaddle
paddleocr
